refactor: improve password recovery flow and add admin cache invalidation

+++ MatHelper.CORE/Models/CountryStatsDto.cs
@@ -3,6 +3,6 @@ namespace MatHelper.CORE.Models
     public class CountryStatsDto
     {
         public required string Country { get; set; }
-        public required ushort Count { get; set; }
+        public required int Count { get; set; }
     }
 }
